mecanicapp/requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.7
django-cleanup==5.0.0
django-import-export==2.3.0
django-userforeignkey==0.3.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.1.2
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
tablib==2.0.0
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
